+++ src/public/js/competition/MatchRunner.js
@@ -1,54 +1,31 @@
 const { remote } = require('electron');
 const fs = require('fs');
 const path = require('path');
-const { spawn } = require('child_process');
+const execa = require('execa');
 const Config = require('./Config');
 
 const { competition, resultFileDirectory } = remote.getCurrentWindow();
 
 class MatchRunner {
   constructor() {
-    this.runningProcesses = [];
+    this.runningProcesses = {};
   }
 
-  runMatch(isBatch = false, switchedSides = false) {
+  runMatch({
+    type = 'single',
+    switchedSides = false,
+    liveEngineStdout = false,
+  } = {}) {
     return new Promise((resolve, reject) => {
       const config = switchedSides ? this.configSwitched : this.config;
 
-      const resultFilePath = path.resolve(resultFileDirectory, `${competition.id}-${isBatch ? 'batch' : 'single'}-resultfile.json`);
+      const resultFilePath = path.resolve(resultFileDirectory, `${competition.id}-${type}-resultfile.json`);
       config.wrapper.resultFile = resultFilePath;
 
-      const proc = spawn('java', ['-jar', competition.paths.matchWrapper, JSON.stringify(config)]);
+      const proc = execa('java', ['-jar', competition.paths.matchWrapper, JSON.stringify(config)]);
 
-      this.runningProcesses.push(proc);
-      const stdout = [];
-
-      proc.stdout.on('data', (data) => {
-        data = data.toString().trim();
-        if (data !== '') stdout.push(data);
-      });
-
-      proc.stderr.on('data', (data) => {
-        data = data.toString().trim();
-        if (data !== '') console.log(data);
-      });
-
-      proc.on('error', (err) => {
-        reject({
-          error: err,
-          stdout: stdout.join('\n'),
-        });
-      });
-
-      proc.on('close', (code) => {
-        this.runningProcesses.splice(this.runningProcesses.indexOf(proc), 1);
-
-        if (code !== 0) {
-          reject({
-            error: new Error(`The match wrapper exited with code ${code}`),
-            stdout: stdout.join('\n'),
-          });
-        }
+      proc.then((result) => {
+        delete this.runningProcesses[type];
 
         const resultFile = JSON.parse(fs.readFileSync(resultFilePath).toString());
 
@@ -57,16 +34,56 @@ class MatchRunner {
 
         const output = {
           resultFile,
-          stdout: stdout.join('\n'),
+          stdout: result.stdout,
         };
 
         if (competition.hasSeed) {
-          output.seed = stdout.join('\n').match(/RANDOM SEED IS: ([0-9a-f-]+)/)[1];
+          output.seed = result.stdout.match(/RANDOM SEED IS: ([0-9a-f-]+)/)[1];
         }
 
         resolve(output);
+      })
+      .catch((error) => {
+        delete this.runningProcesses[type];
+
+        reject({
+          error: new Error(`The match wrapper exited with ${(error.code !== null ? `code ${error.code}` : `signal ${error.signal}`)}`),
+          stdout: error.stdout
+        });
+      });
+
+      this.runningProcesses[type] = proc;
+
+      if (liveEngineStdout) {
+        const $engineTab = $('.log.segment[data-tab="engine-stdout"]');
+        const $engineDimmer = $engineTab.find('.dimmer');
+        const $engineStdout = $engineTab.find('pre');
+
+        proc.stdout.on('data', (data) => {
+          data = data.toString().trim();
+
+          if (data !== '') {
+            $engineDimmer.removeClass('active');
+
+            const text = $engineStdout.text();
+            $engineStdout.text((text + '\n' + data).trim());
+
+            $engineStdout.scrollTop($engineStdout[0].scrollHeight);
+          }
+        });
+      }
+
+      proc.stderr.on('data', (data) => {
+        data = data.toString().trim();
+        if (data !== '') console.log(data);
       });
     });
+  }
+
+  cancel(type) {
+    if (this.runningProcesses.hasOwnProperty(type)) {
+      this.runningProcesses[type].kill();
+    }
   }
 
   updateConfig() {
@@ -82,7 +99,7 @@ class MatchRunner {
   }
 
   exit() {
-    this.runningProcesses.forEach(p => p.kill());
+    Object.keys(this.runningProcesses).forEach(key => this.runningProcesses[key].kill());
   }
 }